+++ src/components/Menus/Sidebar/SettingsPanel/FontSizeSelector/FontSizeSelector.test.js
@@ -22,7 +22,7 @@ describe("When font size cookie unset", () => {
 
   test("Sets cookie to large when first button clicked", async () => {
     const largeButton = fontSelector.getByText(
-      "header.settingsMenu.textSizeOptions.large",
+      "sidebar.settingsMenu.textSizeOptions.large",
     ).parentElement;
     fireEvent.click(largeButton);
     expect(cookies.cookies.fontSize).toBe("large");
@@ -30,7 +30,7 @@ describe("When font size cookie unset", () => {
 
   test("Sets cookie to medium when second button clicked", async () => {
     const mediumButton = fontSelector.getByText(
-      "header.settingsMenu.textSizeOptions.medium",
+      "sidebar.settingsMenu.textSizeOptions.medium",
     ).parentElement;
     fireEvent.click(mediumButton);
     expect(cookies.cookies.fontSize).toBe("medium");
@@ -38,7 +38,7 @@ describe("When font size cookie unset", () => {
 
   test("Sets cookie to small when third button clicked", async () => {
     const smallButton = fontSelector.getByText(
-      "header.settingsMenu.textSizeOptions.small",
+      "sidebar.settingsMenu.textSizeOptions.small",
     ).parentElement;
     fireEvent.click(smallButton);
     expect(cookies.cookies.fontSize).toBe("small");