+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Added
+- Sidebar settings
+
+### Added
 
 - Minor fixes to the sidebar layout
 - LandingPage and ProjectBar's Save buttons use design-system-react